OrcaIRTest is only used in regressions now
